important backend parsing done; need to look at attr-list ONLY; added some basic test cases
