--- feature_scouts/readme.md
@@ -1,4 +1,0 @@
-# Feature proposal
-
- - all the ideas and sample code for a feature goes here
- - use only for code : put descriptions in wiki and not here.
