+++ parser/grammar/grammar.lark
@@ -1,0 +1,104 @@
+start 			: (figure)*
+figure			:"fig" id "{" stmt_list "}"	-> figure_stmt
+
+stmt_list 		: (stmt)*
+stmt 			: size_stmt
+       			| range_stmt
+       			| function_stmt
+       			| plot_stmt 
+       			| print_stmt
+
+
+
+
+size_stmt : "size" "<-" INT [attr_list] -> size_row_stmt
+	| "size" "<-" INT "," INT [attr_list] -> size_row_col_stmt
+	| "size" "<-" "(" INT "," INT ")" [attr_list] -> size_row_col_stmt
+
+
+
+
+range_stmt 	: id "<-" (points | range)	-> range_stmt
+points 	: "[" REAL ("," REAL)* "]"	-> real_points_list
+		| "[" ESC_STR ("," ESC_STR)* "]" -> str_points_list
+
+range : "(" expr "," expr ")" ("step" expr)? -> range_expr
+expr 	: expr OPERATOR expr 	-> expr_op_expr
+		| "(" expr ")"			-> nested_expr
+		| "-" expr 				-> expr_negation
+		| REAL 					-> expr_real
+		| CONST 				-> expr_const
+
+
+
+
+function_stmt 	: id ":=" fn_expr -> func_stmt
+fn_expr : fn_expr OPERATOR fn_expr -> func_op_func
+		| "(" fn_expr ")"	-> nested_func
+		| "-" fn_expr	-> func_negation
+		| id "(" fn_expr ")" -> func_call
+		| REAL -> func_real
+		| id -> func_id
+		| CONST -> func_const
+
+
+
+
+
+
+
+plot_stmt 	: "plot" "(" id "," id ("," [attr_list])? ")"	-> plt_stmt
+
+
+print_stmt		:	"print" "(" id ("," id)* ")" -> print_stmt
+
+attr_list 		: "{" attr ("," attr)* "}"
+attr 			: id "=" ATTR_VALUE		-> attribute
+ATTR_VALUE 		: ESC_STR | REAL | BOOLEAN
+BOOLEAN			: "true" | "false"
+
+
+
+
+
+
+OPERATOR 		: "+" | "-" | "*" | "/" | "^" | "%"
+COMMENT			: 	"#" /.*/
+
+
+
+CONST 	: "PI" | "EXP"
+
+
+
+LCASE_LETTER    : "a".."z"
+UCASE_LETTER    : "A".."Z"
+
+LETTER          : UCASE_LETTER | LCASE_LETTER
+
+id 		        : ID 		-> identifier
+ID 				: ("_"|LETTER) ("_"|LETTER|DIGIT)*
+
+DIGIT           : "0".."9"
+
+INT             : DIGIT+
+SINT      : ["+"|"-"] DIGIT+
+DECIMAL         : INT "." INT? | "." INT
+
+_EXP            : ("e"|"E") SINT
+FLOAT           : INT _EXP | DECIMAL _EXP?
+
+NUMBER          : FLOAT | INT
+REAL		    : ["+"|"-"] NUMBER 
+
+
+ESC_STR 	: ("\"" _STRING_ESC_INNER "\"")
+				| ("'" _STRING_ESC_INNER "'")
+_STRING_INNER       : /.*?/
+_STRING_ESC_INNER   : _STRING_INNER /(?<!\\)(\\\\)*?/
+
+
+WS: /[ \t\f\r\n]/+
+
+%ignore WS
+%ignore COMMENT
