+++ parser/modules/fu_config.py
@@ -1,0 +1,4 @@
+# stores the verbose behavior
+config = {
+	'verbose': False
+}
