+++ parser/modules/fu_keywords.py
@@ -1,0 +1,8 @@
+# the set of keywords available in language
+KEYWORDS = [
+	'fig',
+	'size',
+	'step',
+	'plot',
+	'print'
+]
