--- parser/tests/features/1d-subplot-array_test.fu
@@ -1,19 +1,0 @@
-fig a {
-    size <- 3
-    y <- x 
-    x <- 10 in -0.01 to 0.01 
-    y <- func x
-
-    subplot 0, 2 {
-        dpndnt <- e^(e^indpndnt)
-        indpndnt <- 69 in 1 to 2
-        dpndnt <- func indpndnt
-    }
-
-    subplot 1 {
-        y <- e^i
-        i <- [sqrt(-1)]
-        y <- func i
-    }
-}
-
