+++ tests/features/empty_file.fu
@@ -1,0 +1,1 @@
+# It's perfectly valid for a *.fu file to be empty!
