+++ tests/features/string_test.fu
@@ -1,0 +1,5 @@
+# This script tests the number of ways 
+# we can write strings in WDGAF language
+fig figure{
+	# write your source code here
+}
