+++ tests/old/basic3.fu
