+++ tests/old/test_set/[4]operators.fu
