+++ tests/old/test_set/[5]operators_precedence.fu
