+++ tests/old/test_set/[6]functional_expression.fu
