+++ tests/old/test_set/[7]composed_functions.fu
