+++ tests/old/test_set/[8]functional_list.fu
